[FIX] 순환 참조 문제 해결 #250
